+++ modules/yarpc/thrift.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package rpc
+package yarpc
 
 import (
 	"go.uber.org/fx/service"
